+++ photutils/profiles/curve_of_growth.py
@@ -92,8 +92,10 @@ class CurveOfGrowth(ProfileBase):
     >>> gmodel = Gaussian2D(42.1, 47.8, 52.4, 4.7, 4.7, 0)
     >>> yy, xx = np.mgrid[0:100, 0:100]
     >>> data = gmodel(xx, yy)
-    >>> error = make_noise_image(data.shape, mean=0., stddev=2.4, seed=123)
-    >>> data += error
+    >>> bkg_sig = 2.4
+    >>> noise = make_noise_image(data.shape, mean=0., stddev=bkg_sig, seed=123)
+    >>> data += noise
+    >>> error = np.zeros_like(data) + bkg_sig
 
     Create the curve of growth.
 
@@ -113,16 +115,17 @@ class CurveOfGrowth(ProfileBase):
      5948.92254787 5968.30540534 5931.15611704 5941.94457249 5942.06535486]
 
     >>> print(cog.profile_error)  # doctest: +FLOAT_CMP
-    [  5.32777186   9.37111012  13.41750992  16.62928904  21.7350922
-      25.39862532  30.3867526   34.11478867  39.28263973  43.96047829
-      48.11931395  52.00967328  55.7471834   60.48824739  64.81392778
-      68.71042311  72.71899201  76.54959872  81.33806741  85.98568713
-      91.34841248  95.5173253   99.22190499 102.51980185 106.83601366]
+    [  4.25388924   8.50777848  12.76166773  17.01555697  21.26944621
+      25.52333545  29.7772247   34.03111394  38.28500318  42.53889242
+      46.79278166  51.04667091  55.30056015  59.55444939  63.80833863
+      68.06222787  72.31611712  76.57000636  80.8238956   85.07778484
+      89.33167409  93.58556333  97.83945257 102.09334181 106.34723105]
 
     Plot the curve of growth.
 
     .. plot::
 
+        import matplotlib.pyplot as plt
         import numpy as np
         from astropy.modeling.models import Gaussian2D
 
@@ -134,8 +137,10 @@ class CurveOfGrowth(ProfileBase):
         gmodel = Gaussian2D(42.1, 47.8, 52.4, 4.7, 4.7, 0)
         yy, xx = np.mgrid[0:100, 0:100]
         data = gmodel(xx, yy)
-        error = make_noise_image(data.shape, mean=0., stddev=2.4, seed=123)
-        data += error
+        bkg_sig = 2.4
+        noise = make_noise_image(data.shape, mean=0., stddev=bkg_sig, seed=123)
+        data += noise
+        error = np.zeros_like(data) + bkg_sig
 
         # find the source centroid
         xycen = centroid_quadratic(data, xpeak=48, ypeak=52)
@@ -145,13 +150,15 @@ class CurveOfGrowth(ProfileBase):
         cog = CurveOfGrowth(data, xycen, radii, error=error, mask=None)
 
         # plot the curve of growth
-        cog.plot()
-        cog.plot_error()
+        fig, ax = plt.subplots(figsize=(8, 6))
+        cog.plot(ax=ax)
+        cog.plot_error(ax=ax)
 
     Normalize the profile and plot the normalized curve of growth.
 
     .. plot::
 
+        import matplotlib.pyplot as plt
         import numpy as np
         from astropy.modeling.models import Gaussian2D
 
@@ -163,8 +170,10 @@ class CurveOfGrowth(ProfileBase):
         gmodel = Gaussian2D(42.1, 47.8, 52.4, 4.7, 4.7, 0)
         yy, xx = np.mgrid[0:100, 0:100]
         data = gmodel(xx, yy)
-        error = make_noise_image(data.shape, mean=0., stddev=2.4, seed=123)
-        data += error
+        bkg_sig = 2.4
+        noise = make_noise_image(data.shape, mean=0., stddev=bkg_sig, seed=123)
+        data += noise
+        error = np.zeros_like(data) + bkg_sig
 
         # find the source centroid
         xycen = centroid_quadratic(data, xpeak=48, ypeak=52)
@@ -175,8 +184,9 @@ class CurveOfGrowth(ProfileBase):
 
         # plot the curve of growth
         cog.normalize()
-        cog.plot()
-        cog.plot_error()
+        fig, ax = plt.subplots(figsize=(8, 6))
+        cog.plot(ax=ax)
+        cog.plot_error(ax=ax)
 
     Plot a couple of the apertures on the data.
 
@@ -195,8 +205,10 @@ class CurveOfGrowth(ProfileBase):
         gmodel = Gaussian2D(42.1, 47.8, 52.4, 4.7, 4.7, 0)
         yy, xx = np.mgrid[0:100, 0:100]
         data = gmodel(xx, yy)
-        error = make_noise_image(data.shape, mean=0., stddev=2.4, seed=123)
-        data += error
+        bkg_sig = 2.4
+        noise = make_noise_image(data.shape, mean=0., stddev=bkg_sig, seed=123)
+        data += noise
+        error = np.zeros_like(data) + bkg_sig
 
         # find the source centroid
         xycen = centroid_quadratic(data, xpeak=48, ypeak=52)
@@ -206,11 +218,11 @@ class CurveOfGrowth(ProfileBase):
         cog = CurveOfGrowth(data, xycen, radii, error=error, mask=None)
 
         norm = simple_norm(data, 'sqrt')
-        plt.figure(figsize=(5, 5))
-        plt.imshow(data, norm=norm, origin='lower')
-        cog.apertures[5].plot(color='C0', lw=2)
-        cog.apertures[10].plot(color='C1', lw=2)
-        cog.apertures[15].plot(color='C3', lw=2)
+        fig, ax = plt.subplots(figsize=(5, 5))
+        ax.imshow(data, norm=norm, origin='lower')
+        cog.apertures[5].plot(ax=ax, color='C0', lw=2)
+        cog.apertures[10].plot(ax=ax, color='C1', lw=2)
+        cog.apertures[15].plot(ax=ax, color='C3', lw=2)
     """
 
     def __init__(self, data, xycen, radii, *, error=None, mask=None,